+++ modules/diffusion/diffusion_solver.h
@@ -5,7 +5,7 @@
 
 #include "modules/diffusion/boundary.h"
 #include "framework/object_factory.h"
-#include "framework/physics/solver_base/solver.h"
+#include "framework/physics/solver.h"
 #include "framework/math/petsc_utils/petsc_utils.h"
 
 namespace opensn
