+++ modules/point_reactor_kinetics/point_reactor_kinetics.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "framework/physics/solver_base/solver.h"
+#include "framework/physics/solver.h"
 #include "framework/math/math.h"
 #include "framework/math/dynamic_matrix.h"
 #include "framework/math/dynamic_vector.h"
